+++ src/test/compile-fail/borrowck-overloaded-index-2.rs
@@ -16,10 +16,10 @@ struct MyVec<T> {
     data: Vec<T>,
 }
 
-impl<T> Index<uint> for MyVec<T> {
+impl<T> Index<usize> for MyVec<T> {
     type Output = T;
 
-    fn index(&self, &i: &uint) -> &T {
+    fn index(&self, &i: &usize) -> &T {
         &self.data[i]
     }
 }
