+++ src/test/compile-fail/mutable-class-fields.rs
@@ -9,11 +9,11 @@
 // except according to those terms.
 
 struct cat {
-  meows : uint,
+  meows : usize,
   how_hungry : isize,
 }
 
-fn cat(in_x : uint, in_y : isize) -> cat {
+fn cat(in_x : usize, in_y : isize) -> cat {
     cat {
         meows: in_x,
         how_hungry: in_y
