+++ src/test/compile-fail/unboxed-closure-sugar-used-on-struct-3.rs
@@ -19,7 +19,7 @@ impl<A,B> Bar<A,B> {
 }
 
 fn bar() {
-    let b = Box::Bar::<isize,uint>::new(); // OK
+    let b = Box::Bar::<isize,usize>::new(); // OK
 
     let b = Box::Bar::()::new();
     //~^ ERROR expected ident, found `(`
